Bottle 스프라이트를  프리펩으로 만들기 적합하게 수정
