+++ Assets/Scenes/Stage.unity
@@ -1724,7 +1724,7 @@ MonoBehaviour:
   isActive: 0
   trajectoryDotPrefab: {fileID: 1152180298276607188, guid: 76393c19bc2f0f8439169d036e64bbc5,
     type: 3}
-  trajectoryNumber: 10
+  trajectoryNumber: 15
 --- !u!224 &1914826069
 RectTransform:
   m_ObjectHideFlags: 0
