+++ Assets/Scripts/PadStrength.cs
@@ -9,67 +9,22 @@ using UnityEngine.UI;
 */
 public class PadStrength : MonoBehaviour, IPointerDownHandler, IPointerUpHandler
 {
-
-
-    public Bottle bottle;
+    public BottleController bottleController;
     public bool isTouch = false;
-    private bool isThrowing = false; //캐릭터가 물병을 던지는 동작을 진행중인가의 여부
-    private float delayTime = 1f; //힘 조절 버튼에서 손가락을 때고 물병이 던져지기까지의 딜레이 타임
-    public int addStrength = 7; //시간별로 더해지는 힘 값, 조정 가능
-
-    //{get;set;}을 하면 코드 내에서 수정은 가능하나, 유니티에서 보여지지 않음
-    public float totalStrength { get; set; }
-
-    GameObject strengthGauge; //힘 게이지 스프라이트
-
 
     void Start()
     {
-        this.strengthGauge = GameObject.Find("StrengthGauge");
+        bottleController = GameObject.Find("Bottle").GetComponent<BottleController>();
     }
 
-
-    void Update()
-    {                                                                     
-        if (isTouch && (!bottle.isLaunched) && (!isThrowing) && bottle.isActive && (totalStrength < 2 * addStrength))//패드가 눌려있고 물병을 던지는 도중이 아니며 물병이 날아가는 도중이 아니면
-        {
-
-            totalStrength += addStrength*Time.deltaTime; // 매 초마다 일정한 힘을 더한다.
-            this.strengthGauge.GetComponent<Image>().fillAmount += 1.0f / 2 * Time.deltaTime; // 매 초마다 힘 게이지가 1/2 씩 차오른다.
-        }
-
-        if (isThrowing || (totalStrength >= 2*addStrength)) //패드에서 마우스를 뗐거나 힘 버튼을 2초 이상 눌렀을 때
-        {
-            delayTime -= Time.fixedDeltaTime; //딜레이 타임만큼 던지는 동작이 지연된다.
-
-            if (delayTime <= 0)
-            {
-                strengthGauge.gameObject.SetActive(false); // 힘 게이지를 화면에서 제거한다.
-                bottle.Jump();
-                isThrowing = false;
-                totalStrength = 0;
-            }
-        }
-    }
-
-   
     public void OnPointerDown(PointerEventData eventData) //패드를 누르는 순간
     {
-        if ((!bottle.isLaunched) && (!isThrowing) && bottle.isActive)
-        {
-            delayTime = 1f; //딜레이 타임 초기화
-            isTouch = true;
-            strengthGauge.gameObject.SetActive(true); //힘 게이지를 화면에 표시한다.
-        }
+        bottleController.StrengthButtonDown();
     }
 
     public void OnPointerUp(PointerEventData eventData) //패드에서 마우스를 떼는 순간
     {
-        if (bottle.isActive)
-        {
-            isTouch = false;
-            if (bottle.isLaunched) isThrowing = false;
-            else isThrowing = true;
-        }
+        bottleController.StrengthButtonUp();
     }
+
 }
